fix: moved image folder into src
